+++ src/main/java/com/prgms/allen/dining/domain/reservation/entity/VisitStatus.java
@@ -1,8 +1,6 @@
-package com.prgms.allen.dining.domain.reservation.dto;
+package com.prgms.allen.dining.domain.reservation.entity;
 
 import java.util.List;
-
-import com.prgms.allen.dining.domain.reservation.entity.ReservationStatus;
 
 public enum VisitStatus {
