+++ src/test/java/com/prgms/allen/dining/web/domain/owner/reservation/OwnerReservationApiTest.java
@@ -14,11 +14,14 @@ import java.time.temporal.ChronoUnit;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
+import org.springframework.restdocs.payload.JsonFieldType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -30,6 +33,7 @@ import com.prgms.allen.dining.domain.reservation.dto.ReservationStatusUpdateReq;
 import com.prgms.allen.dining.domain.reservation.entity.FakeReservationCustomerInput;
 import com.prgms.allen.dining.domain.reservation.entity.Reservation;
 import com.prgms.allen.dining.domain.reservation.entity.ReservationCustomerInput;
+import com.prgms.allen.dining.domain.reservation.entity.ReservationStatus;
 import com.prgms.allen.dining.domain.restaurant.RestaurantRepository;
 import com.prgms.allen.dining.domain.restaurant.entity.Restaurant;
 import com.prgms.allen.dining.generator.DummyGenerator;
@@ -197,6 +201,70 @@ class OwnerReservationApiTest {
 				requestFields(
 					fieldWithPath("status").description("변경할 상태")
 				))
+			);
+	}
+
+	@ParameterizedTest
+	@CsvSource({"PENDING", "CONFIRMED", "VISITED", "CANCELLED", "NO_SHOW"})
+	@DisplayName("점주는 식당의 예약 목록을 조회할 수 있다.")
+	void getOwnerReservations(String status) throws Exception {
+		// given
+		Member customer = memberRepository.save(DummyGenerator.CUSTOMER);
+		Member owner = memberRepository.save(DummyGenerator.OWNER);
+		Restaurant restaurant = restaurantRepository.save(DummyGenerator.createRestaurant(owner));
+		ReservationCustomerInput customerInput = new FakeReservationCustomerInput(
+			LocalDate.now()
+				.minusDays(1),
+			LocalTime.now()
+				.truncatedTo(ChronoUnit.HOURS),
+			2
+		);
+		Reservation reservation = reservationRepository.save(
+			Reservation.newTestInstance(null, customer, restaurant, ReservationStatus.valueOf(status), customerInput)
+		);
+		String page = "0";
+		String size = "5";
+
+		mockMvc.perform(get("/owner/api/reservations")
+				.param("reservationStatus", status)
+				.param("restaurantId", restaurant.getId().toString())
+				.param("page", page)
+				.param("size", size)
+				.contentType(MediaType.APPLICATION_JSON))
+			.andExpect(status().isOk())
+			.andDo(print())
+			.andDo(
+				document("reservation-getAllByCustomer",
+					requestParameters(
+						parameterWithName("reservationStatus").description("조회할 상태"),
+						parameterWithName("restaurantId").description("조회하려는 레스토랑의 아이디"),
+						parameterWithName("page").description("페이지 번호"),
+						parameterWithName("size").description("한 페이지 당 조회 개수")
+					),
+					responseFields(
+						fieldWithPath("content[]").type(JsonFieldType.ARRAY).description("조회된 예약들"),
+						fieldWithPath("content[].visitorName").type(JsonFieldType.STRING).description("방문자 이름"),
+						fieldWithPath("content[].phone").type(JsonFieldType.STRING).description("방문자 핸드폰 번호"),
+						fieldWithPath("content[].visitDateTime").type(JsonFieldType.STRING).description("예약한 시간 및 날짜"),
+						fieldWithPath("content[].visitorCount").type(JsonFieldType.NUMBER).description("방문 인원수"),
+
+						fieldWithPath("pageable").type(JsonFieldType.STRING).description(""),
+						fieldWithPath("totalElements").type(JsonFieldType.NUMBER).description("전체 데이터 개수"),
+						fieldWithPath("first").type(JsonFieldType.BOOLEAN).description("첫번째 페이지인지 여부"),
+						fieldWithPath("last").type(JsonFieldType.BOOLEAN).description("마지막 페이지인지 여부"),
+						fieldWithPath("totalPages").type(JsonFieldType.NUMBER).description("전체 페이지 개수"),
+						fieldWithPath("numberOfElements").type(JsonFieldType.NUMBER).description("요청 페이지에서 조회된 데이터 개수"),
+						fieldWithPath("size").type(JsonFieldType.NUMBER).description("한 페이지당 조회할 데이터 개수"),
+						fieldWithPath("number").type(JsonFieldType.NUMBER).description("현재 페이지 번호"),
+
+						fieldWithPath("sort").type(JsonFieldType.OBJECT).description("정렬 기준"),
+						fieldWithPath("sort.sorted").type(JsonFieldType.BOOLEAN).description("정렬 됐는지 여부"),
+						fieldWithPath("sort.unsorted").type(JsonFieldType.BOOLEAN).description("정렬 안됐는지 여부"),
+						fieldWithPath("sort.empty").type(JsonFieldType.BOOLEAN).description("데이터가 비었는지 여부"),
+
+						fieldWithPath("empty").type(JsonFieldType.BOOLEAN).description("데이터가 비었는지 여부")
+					)
+				)
 			);
 	}
 }
